+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add export to SVG for diagrams https://github.com/opencaesar/oml-vision/pull/3
 - Add VSCode dev container https://github.com/opencaesar/oml-vision/pull/21
 - Add ability to change node and node text color while automatically changing node color based on node type https://github.com/opencaesar/oml-vision/pull/22
-- Add ability to change edge color and determine if edge is animated https://github.com/opencaesar/oml-vision/pull/23
+- Add ability to animate edges https://github.com/opencaesar/oml-vision/pull/23
 
 ## [0.1.0]