+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add VSCode dev container https://github.com/opencaesar/oml-vision/pull/21
 - Add ability to change node and node text color while automatically changing node color based on node type https://github.com/opencaesar/oml-vision/pull/22
 - Add ability to animate edges https://github.com/opencaesar/oml-vision/pull/23
+- Add ability to highlight/unlight connected edges to selected nodes https://github.com/opencaesar/oml-vision/pull/24
 
 ## [0.1.0]